Add enum casing handling
Also refactored tests to fix these problems:
 * most of the steps were in one massive class
 * lots of things were going through the black hold that is the ScenarioContext
 * the use of the per-ScenarioContext DI container wasn't working because tests needed to modify the DI setup
