--- Solutions/Corvus.Json.Specs/JsonExtensionsContainerBindings.cs
@@ -1,38 +1,0 @@
-﻿// <copyright file="JsonExtensionsContainerBindings.cs" company="Endjin Limited">
-// Copyright (c) Endjin Limited. All rights reserved.
-// </copyright>
-
-namespace Corvus.Json.Specs;
-
-using System.Text.Json;
-using System.Text.Json.Serialization;
-
-using Corvus.Testing.SpecFlow;
-
-using Microsoft.Extensions.DependencyInjection;
-
-using TechTalk.SpecFlow;
-
-/// <summary>
-/// Provides Specflow bindings for Endjin Composition.
-/// </summary>
-[Binding]
-public static class JsonExtensionsContainerBindings
-{
-    /// <summary>
-    /// Setup the endjin container for a feature.
-    /// </summary>
-    /// <remarks>We expect features run in parallel to be executing in separate app domains.</remarks>
-    /// <param name="featureContext">The SpecFlow test context.</param>
-    [BeforeFeature("@perFeatureContainer", Order = ContainerBeforeFeatureOrder.PopulateServiceCollection)]
-    public static void SetupFeature(FeatureContext featureContext)
-    {
-        ContainerBindings.ConfigureServices(
-            featureContext,
-            serviceCollection => serviceCollection
-                .AddJsonSerializerOptionsProvider()
-                .AddJsonPropertyBagFactory()
-                .AddJsonCultureInfoConverter()
-                .AddSingleton<JsonConverter>(new JsonStringEnumConverter(JsonNamingPolicy.CamelCase)));
-    }
-}
