+++ Solutions/Corvus.Json.Specs/Samples/ExampleEnum2.cs
@@ -1,0 +1,12 @@
+﻿// <copyright file="ExampleEnum2.cs" company="Endjin Limited">
+// Copyright (c) Endjin Limited. All rights reserved.
+// </copyright>
+
+namespace Corvus.Extensions.Json.Specs.Samples;
+
+public enum ExampleEnum2
+{
+    ValueOne,
+    ValueTwo,
+    ValueThree,
+}
